+++ CHANGELOG.md
@@ -14,6 +14,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Note!!!
 # After update to v7.5.0 and above from plevious version the sensors in plugin config need to be configured again!!!
 
+## [7.7.2] - (26.11.2023)
+## Changes
+- added debug log for services prepare
+- prepare for generators support
+- cleanup
+
 ## [7.7.1] - (26.11.2023)
 ## Changes
 - dynamically add *Live Data* characteristics based on installed devices